fix(resources): move ResourceCreationError to separate file
Fixes TypeScript error with 'use server' directive by moving the error class to its own file, as Server Components only allow async functions to be exported from server files.
